chore: modularized code with hooks
