+++ src/types/api.ts
@@ -1,0 +1,15 @@
+export interface IData {
+  id: string;
+  Date: string;
+  Company: string;
+  Title: string;
+  Location: string;
+  Salary: number;
+  Description: string;
+  Level: string;
+  Type: string;
+  Function: string;
+  Industry: string;
+  Link: string;
+  image_url: string;
+}
